+++ packages/mobx-binder-core/src/conversion/EmptyStringConverter.ts
@@ -1,7 +1,12 @@
 import { Converter } from './Converter'
+import { createLabel } from '../validation/Labels'
 
 export class EmptyStringConverter<X> implements Converter<any, string, string | X> {
-    constructor(private targetValue: X) {}
+    readonly label: string
+
+    constructor(private targetValue: X) {
+        this.label = createLabel('EmptyStringConverter', { targetValue })
+    }
 
     public convertToModel(value: string): string | X {
         return value === '' ? this.targetValue : value