+++ packages/mobx-binder-core/src/model/binder/chain/AsyncConvertingModifier.ts
@@ -3,7 +3,7 @@ import { Data, Modifier, SyncValueValidationResult, ValidValueValidationResult }
 import { Context } from '../Context'
 import { AbstractModifier } from './AbstractModifier'
 import { AsyncConverter } from '../../../conversion/Converter'
-import { isValidationError } from '../../../conversion/ValidationError'
+import { isValidationError, ValidationError } from '../../../conversion/ValidationError'
 import { Validity } from '../../../validation/Validity'
 
 type AsyncConversionInfo<ViewType, ModelType, ValidationResult> =
@@ -52,6 +52,15 @@ export class AsyncConvertingModifier<ValidationResult, ViewType, ModelType> exte
             validity: computed,
             startNewConversion: action,
         })
+    }
+
+    get name() {
+        const name = this.converter.label || this.converter.constructor.name
+        return name && name !== 'Function' ? name : undefined
+    }
+
+    get type() {
+        return 'async conversion'
     }
 
     get data(): Data<ModelType> {
@@ -145,7 +154,7 @@ export class AsyncConvertingModifier<ValidationResult, ViewType, ModelType> exte
                         this.info = {
                             status: 'invalid',
                             validatedValue: value,
-                            lastValidationResult: err.validationResult,
+                            lastValidationResult: (err as ValidationError<ValidationResult>).validationResult,
                         }
                     }
                 })