+++ packages/mobx-binder-core/src/model/fields/AbstractField.ts
@@ -1,44 +1,65 @@
-import { action, autorun, makeObservable, observable, runInAction } from 'mobx'
+import { action, makeObservable, observable } from 'mobx'
 import { FieldStore } from './FieldStore'
+import { Binding } from '../binder/Binder'
+import sleep from '../../utils/sleep'
 
 export abstract class AbstractField<ValueType> implements FieldStore<ValueType> {
     public readOnly = false
-
-    public required = false
-
-    public changed = false
 
     public showValidationResults = false
 
     public abstract value: ValueType
 
-    public valid?: boolean = undefined
-
-    public validating = false
-
     public touched = false
 
     public visited = false
 
-    public errorMessage?: string = undefined
+    private binding?: Binding<unknown, unknown> = undefined
 
     protected constructor(public readonly valueType: string, public readonly name: string) {
         makeObservable(this, {
-            readOnly: observable,
-            required: observable,
-            changed: observable,
             showValidationResults: observable,
-            valid: observable,
-            validating: observable,
             touched: observable,
             visited: observable,
-            errorMessage: observable,
-
-            updateValue: action,
-            handleFocus: action,
-            handleBlur: action,
-            reset: action,
+            updateValue: action.bound,
+            handleFocus: action.bound,
+            handleBlur: action.bound,
+            reset: action.bound,
         })
+    }
+
+    get valid() {
+        this.assertBound(this.binding)
+        return this.binding.valid
+    }
+
+    get required() {
+        this.assertBound(this.binding)
+        return this.binding.required
+    }
+
+    get validating() {
+        this.assertBound(this.binding)
+        return this.binding.validating
+    }
+
+    get errorMessage() {
+        this.assertBound(this.binding)
+        return this.binding.errorMessage
+    }
+
+    set errorMessage(customErrorMessage: string | undefined) {
+        this.assertBound(this.binding)
+        this.binding.customErrorMessage = customErrorMessage
+    }
+
+    get changed() {
+        this.assertBound(this.binding)
+        return this.binding.changed
+    }
+
+    public bind(binding: Binding<unknown, unknown>) {
+        this.binding = binding
     }
 
     public updateValue(newValue: ValueType) {
@@ -51,22 +72,26 @@ export abstract class AbstractField<ValueType> implements FieldStore<ValueType> 
     }
 
     public handleBlur(): void {
-        autorun(
-            () =>
-                runInAction(() => {
-                    this.showValidationResults = true
-                }),
-            { name: 'handleBlur', delay: 100 },
+        void Promise.all([sleep(100), this.binding?.validateOnBlur()]).then(
+            action(() => {
+                this.showValidationResults = true
+            }),
         )
     }
 
     public reset(value: ValueType) {
+        this.assertBound(this.binding)
         this.value = value
-        this.changed = false
+        this.binding.setUnchanged()
+        this.binding.customErrorMessage = undefined
         this.touched = false
         this.visited = false
         this.showValidationResults = false
-        this.valid = undefined
-        this.errorMessage = undefined
+    }
+
+    private assertBound(binding?: Binding<unknown, unknown>): asserts binding {
+        if (!binding) {
+            throw new Error('Trying to use an unbound field')
+        }
     }
 }