+++ packages/sample/.eslintrc.js
@@ -2,4 +2,7 @@ module.exports =  {
     extends:  [
         '../mobx-binder-utils/.eslintrc.js',
     ],
+    parserOptions: {
+        project: 'tsconfig.json'
+    },
 };